requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.4.2

# Machine learning and deep learning
torch==2.2.1
numpy==1.24.3
scikit-learn==1.3.2
sentence-transformers==2.2.2

# Power system analysis
opendss-direct==0.1.0
pandapower==2.14.1

# Template engine
Jinja2==3.1.2

# Utilities
python-multipart==0.0.6
aiofiles==23.2.1
tqdm==4.66.1
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
black==23.10.1
isort==5.12.0
flake8==6.1.0
mypy==1.6.1